fix MultipleOptionsSearchBarCmp click option is missing the context
clicking on a option leads to search always on the first possible
option. now the context of the click is given to the handle function,
similar to the click handle function in the MultipleOptionsSearchBar
component. The context (in this function called "result") gives then
the correct destinationURL.

NOTE: bring back the commented default option